integrate Prompt component for react-location
