+++ packages/react-location/src/routerComponent.tsx
@@ -1,8 +1,5 @@
 import React from "react";
 import {
-    ErrorComponent,
-    LayoutWrapper,
-    ResourceRouterParams,
     useAuthenticated,
     useIsAuthenticated,
     useRefineContext,
@@ -12,9 +9,7 @@ import {
 } from "@pankod/refine";
 import { rankRoutes } from "react-location-rank-routes";
 
-import { BrowserRouter, BrowserRouterProps } from "react-router-dom";
 import {
-    createHashHistory,
     Router,
     ReactLocation,
     Route,
@@ -22,30 +17,13 @@ import {
     useRouter,
 } from "react-location";
 
-import { RouteProvider } from "./routeProvider";
+import { ResourceComponentWrapper } from "./resourceComponent";
 
-const hashHistory = createHashHistory();
 const location = new ReactLocation();
 
-export const RouterComponent: React.FC<BrowserRouterProps> = ({
-    children,
-    ...props
-}) => {
-    /*     const { setWarnWhen } = useWarnAboutChange();
-
-    const getUserConfirmation: (
-        message: string,
-        callback: (ok: boolean) => void,
-    ) => void = (message, callback) => {
-        const allowTransition = window.confirm(message);
-        if (allowTransition) {
-            setWarnWhen(false);
-        }
-        callback(allowTransition);
-    }; */
-
+export const RouterComponent: React.FC = () => {
     const { resources } = useResource();
-    const { catchAll, DashboardPage, LoginPage } = useRefineContext();
+    const { DashboardPage, LoginPage } = useRefineContext();
 
     const { routes: customRoutes }: { routes: Route[] } = useRouterContext();
 
@@ -71,7 +49,7 @@ export const RouterComponent: React.FC<BrowserRouterProps> = ({
             },
             {
                 path: "*",
-                element: <LoginWithNavigate />,
+                element: <LoginNavigateWithToParam />,
             },
         ];
     }
@@ -122,108 +100,13 @@ export const RouterComponent: React.FC<BrowserRouterProps> = ({
         ];
     }
 
-    console.log("customRoutes", customRoutes, routes);
-
     return (
         <Router location={location} routes={routes} filterRoutes={rankRoutes} />
     );
 };
 
-export const ResourceComponentWrapper: React.FC = () => {
-    const { catchAll } = useRefineContext();
-    const { useParams } = useRouterContext();
-    const { resources } = useResource();
-
-    const { resource: routeResourceName, action } =
-        useParams<ResourceRouterParams>();
-
-    const resource = resources.find((res) => res.route === routeResourceName);
-
-    if (resource) {
-        const {
-            list,
-            create,
-            edit,
-            show,
-            name,
-            canCreate,
-            canEdit,
-            canShow,
-            canDelete,
-        } = resource;
-
-        const List = list ?? (() => null);
-        const Create = create ?? (() => null);
-        const Edit = edit ?? (() => null);
-        const Show = show ?? (() => null);
-
-        const renderCrud = () => {
-            switch (action) {
-                case undefined: {
-                    return (
-                        <List
-                            name={name}
-                            canCreate={canCreate}
-                            canEdit={canEdit}
-                            canDelete={canDelete}
-                            canShow={canShow}
-                        />
-                    );
-                }
-
-                case "create": {
-                    return (
-                        <Create
-                            name={name}
-                            canCreate={canCreate}
-                            canEdit={canEdit}
-                            canDelete={canDelete}
-                            canShow={canShow}
-                        />
-                    );
-                }
-
-                case "edit": {
-                    return (
-                        <Edit
-                            name={name}
-                            canCreate={canCreate}
-                            canEdit={canEdit}
-                            canDelete={canDelete}
-                            canShow={canShow}
-                        />
-                    );
-                }
-
-                case "show": {
-                    return (
-                        <Show
-                            name={name}
-                            canCreate={canCreate}
-                            canEdit={canEdit}
-                            canDelete={canDelete}
-                            canShow={canShow}
-                        />
-                    );
-                }
-            }
-        };
-
-        return <LayoutWrapper>{renderCrud()}</LayoutWrapper>;
-    }
-
-    return catchAll ? (
-        <>{catchAll}</>
-    ) : (
-        <LayoutWrapper>
-            <ErrorComponent />
-        </LayoutWrapper>
-    );
-};
-
-export const LoginWithNavigate: React.FC = () => {
+export const LoginNavigateWithToParam: React.FC = () => {
     const { state } = useRouter();
-    console.log("location", location);
 
     return (
         <Navigate
